re-organise get_changed_files
