+++ apps/wizard/app_pages/indicator_upgrade/utils.py
@@ -7,8 +7,9 @@ from pymysql import OperationalError
 from rapidfuzz import fuzz
 from structlog import get_logger
 
-from apps.utils.map_datasets import get_changed_files, get_grapher_changes
+from apps.utils.map_datasets import get_grapher_changes
 from etl.db import config, get_all_datasets, get_connection, get_variables_in_dataset
+from etl.git import get_changed_files
 from etl.indicator_upgrade.schema import get_schema_chart_config
 from etl.version_tracker import VersionTracker
